6. Testing coroutines through behavior
https://kotlin.github.io/kotlinx.coroutines/kotlinx-coroutines-test/
testImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.4.2'

By calling onMainViewClicked, the coroutine we just created will be launched. This test checks that the taps text stays "0 taps" right after onMainViewClicked is called, then 1 second later it gets updated to "1 taps".

This test uses virtual-time to control the execution of the coroutine launched by onMainViewClicked. The MainCoroutineScopeRule lets you pause, resume, or control the execution of coroutines that are launched on the Dispatchers.Main. Here we're calling advanceTimeBy(1_000) which will cause the main dispatcher to immediately execute coroutines that are scheduled to resume 1 second later.