6. Building logic with LiveData
The LiveData builder allows us to calculate values asynchronously, as liveData is backed by coroutines.
Here we have a suspend function to fetch a LiveData list of plants from the database, while also calling a suspend function to get the custom sort order.
We then combine these two values to sort the list of plants and return the value, all within the builder.

You can emit multiple values from a LiveData by calling the emitSource() function whenever you want to emit a new value. Note that each call to emitSource() removes the previously-added source.

If any of the suspend function calls fail, the entire block is canceled and not restarted, which helps avoid leaks.